refactor: extract app review actions
